Request-Discriminator to Request-Tag rephrasing (part 1)
This is the first part of the rephrasing into being a block extension
rather than a security-context extension as was the outcome of the last
OSCOAP meeting. (Rationale is that it's easier both to reason about a
Request-Tag and to make it understood; if it's as equivalent to a
subchannel as I think it is, that can still later be shown when the
problem is better undstood).

Abstract and introduction were completely rewritten, and the problem
set in the introduction. The topic of usage outside OSCOAP (typically
for outer-blockwise OSCOAP messages by proxies) was dropped from the
rationale to be later added in an appendix.

This adds a section of reflection on the various combination of block
options.